requirements.txt
pyexiftool
python-ffmpeg